code reorg and cleanup
