+++ refined-types-plugin/src/main/kotlin/arrow/meta/plugins/liquid/LiquidPlugin.kt
@@ -6,7 +6,7 @@ import arrow.meta.invoke
 import arrow.meta.plugins.liquid.phases.ir.registerIrInterpreterCompileTimeFunctions
 import arrow.meta.plugins.liquid.phases.ir.validateIrCallsToRefinedFunctions
 import arrow.meta.plugins.liquid.phases.quotes.generateRefinedApi
-import arrow.meta.plugins.liquid.phases.solver.state.solverStateAnalysis
+import arrow.meta.plugins.liquid.phases.analysis.solverStateAnalysis
 
 
 val Meta.liquidExpressions: CliPlugin
