+++ refined-types-plugin/src/main/kotlin/arrow/meta/plugins/liquid/phases/analysis/solver/ConstraintChecks.kt
@@ -1,170 +1,37 @@
-package arrow.meta.plugins.liquid.phases.solver.state
+package arrow.meta.plugins.liquid.phases.analysis.solver
 
-import arrow.meta.Meta
-import arrow.meta.continuations.*
+import arrow.meta.continuations.Cont
+import arrow.meta.continuations.SimpleCont
+import arrow.meta.continuations.contEach
+import arrow.meta.continuations.continueWith
+import arrow.meta.continuations.forget
+import arrow.meta.continuations.guard
+import arrow.meta.continuations.orDoNothing
+import arrow.meta.continuations.orElse
+import arrow.meta.continuations.runCont
+import arrow.meta.continuations.then
+import arrow.meta.continuations.wrap
 import arrow.meta.phases.CompilerContext
-import arrow.meta.phases.Composite
-import arrow.meta.phases.ExtensionPhase
 import arrow.meta.phases.analysis.body
-import arrow.meta.plugins.liquid.errors.MetaErrors.*
-import arrow.meta.plugins.liquid.phases.solver.Solver
-import arrow.meta.plugins.liquid.phases.solver.collector.*
-import arrow.meta.plugins.liquid.phases.solver.collector.constraints
-import arrow.meta.plugins.liquid.phases.solver.collector.postCall
-import arrow.meta.plugins.liquid.phases.solver.collector.preCall
-import org.jetbrains.kotlin.analyzer.AnalysisResult
+import arrow.meta.plugins.liquid.errors.MetaErrors
+import arrow.meta.plugins.liquid.phases.solver.collector.renameDeclarationConstraints
 import org.jetbrains.kotlin.codegen.kotlinType
-import org.jetbrains.kotlin.descriptors.CallableDescriptor
 import org.jetbrains.kotlin.descriptors.DeclarationDescriptor
-import org.jetbrains.kotlin.descriptors.ModuleDescriptor
 import org.jetbrains.kotlin.fir.lightTree.converter.nameAsSafeName
-import org.jetbrains.kotlin.psi.*
-import org.jetbrains.kotlin.psi.psiUtil.getChildrenOfType
-import org.jetbrains.kotlin.resolve.BindingTrace
+import org.jetbrains.kotlin.psi.KtBlockExpression
+import org.jetbrains.kotlin.psi.KtCallExpression
+import org.jetbrains.kotlin.psi.KtConstantExpression
+import org.jetbrains.kotlin.psi.KtDeclaration
+import org.jetbrains.kotlin.psi.KtDeclarationWithBody
+import org.jetbrains.kotlin.psi.KtDeclarationWithInitializer
+import org.jetbrains.kotlin.psi.KtExpression
+import org.jetbrains.kotlin.psi.KtNamedDeclaration
+import org.jetbrains.kotlin.psi.KtSimpleNameExpression
+import org.jetbrains.kotlin.psi.KtVariableDeclaration
 import org.jetbrains.kotlin.resolve.calls.callUtil.getResolvedCall
-import org.jetbrains.kotlin.resolve.calls.model.ResolvedCall
 import org.jetbrains.kotlin.resolve.checkers.DeclarationCheckerContext
-import org.jetbrains.kotlin.resolve.descriptorUtil.fqNameSafe
 import org.jetbrains.kotlin.types.typeUtil.isInt
-import org.sosy_lab.java_smt.api.*
-import java.util.concurrent.atomic.AtomicReference
-
-data class SolverState(
-  val solver: Solver = Solver(),
-  val prover: ProverEnvironment = solver.newProverEnvironment(
-    SolverContext.ProverOptions.GENERATE_MODELS,
-    SolverContext.ProverOptions.GENERATE_UNSAT_CORE),
-  val callableConstraints: MutableList<DeclarationConstraints> = mutableListOf(),
-  val names: NameProvider = NameProvider()
-) {
-
-  private var stage = Stage.Init
-
-  fun currentStage(): Stage = stage
-
-  fun collecting(): Unit {
-    stage = Stage.CollectConstraints
-  }
-
-  fun collectionEnds(): Unit {
-    stage = Stage.Prove
-  }
-
-  fun <A> bracket(f: () -> A): A {
-    prover.push()
-    val result = f()
-    prover.pop()
-    return result
-  }
-
-  fun isIn(that: Stage) = stage == that
-
-  companion object {
-
-    fun key(moduleDescriptor: ModuleDescriptor): String =
-      "SolverState-${moduleDescriptor.name}"
-  }
-
-  enum class Stage {
-    Init, CollectConstraints, Prove
-  }
-}
-
-class NameProvider {
-  private val counter = AtomicReference(0)
-
-  fun newName(prefix: String): String {
-    val n = counter.getAndUpdate { it + 1 }
-    return "${prefix}${n}"
-  }
-}
-
-internal fun Meta.solverStateAnalysis(): ExtensionPhase =
-  Composite(
-    listOf(
-      analysis(
-        doAnalysis = { project, module, projectContext, files, bindingTrace, componentProvider ->
-          ensureSolverStateInitialization(module)
-          null
-        },
-        analysisCompleted = { project, module, bindingTrace, files ->
-          finalizeConstraintsCollection(module, bindingTrace)
-        },
-      ),
-      declarationChecker { declaration, descriptor, context ->
-        collectDeclarationsConstraints(context, declaration, descriptor)
-      },
-      declarationChecker { declaration, descriptor, context ->
-        checkDeclarationConstraints(context, declaration, descriptor)
-      }
-    )
-  )
-
-// PHASE 1: COLLECTION OF CONSTRAINTS
-// ==================================
-
-internal fun SolverState.constraintsFromSolverState(resolvedCall: ResolvedCall<*>): DeclarationConstraints? =
-  callableConstraints.firstOrNull {
-    resolvedCall.resultingDescriptor.fqNameSafe == it.descriptor.fqNameSafe
-  }
-
-internal fun SolverState.constraintsFromSolverState(descriptor: DeclarationDescriptor): DeclarationConstraints? =
-  callableConstraints.firstOrNull {
-    descriptor.fqNameSafe == it.descriptor.fqNameSafe
-  }
-
-private fun CompilerContext.collectDeclarationsConstraints(
-  context: DeclarationCheckerContext,
-  declaration: KtDeclaration,
-  descriptor: DeclarationDescriptor
-) {
-  val solverState = get<SolverState>(SolverState.key(context.moduleDescriptor))
-  if (solverState != null && (solverState.isIn(SolverState.Stage.Init) || solverState.isIn(SolverState.Stage.CollectConstraints))) {
-    solverState.collecting()
-    val solver = solverState.solver
-    val constraints = declaration.constraints(solver, context.trace.bindingContext)
-    solverState.addConstraintsToSolverState(constraints, descriptor, declaration)
-  }
-}
-
-private fun SolverState.addConstraintsToSolverState(
-  constraints: List<Pair<ResolvedCall<*>, BooleanFormula>>,
-  descriptor: DeclarationDescriptor,
-  declaration: KtDeclaration
-) {
-  if (constraints.isNotEmpty()) {
-    val preConstraints = arrayListOf<BooleanFormula>()
-    val postConstraints = arrayListOf<BooleanFormula>()
-    constraints.forEach { (call, formula) ->
-      if (call.preCall()) preConstraints.add(formula)
-      if (call.postCall()) postConstraints.add(formula)
-    }
-    callableConstraints.add(
-      DeclarationConstraints(descriptor, declaration, preConstraints, postConstraints)
-    )
-  }
-}
-
-private fun CompilerContext.finalizeConstraintsCollection(
-  module: ModuleDescriptor,
-  bindingTrace: BindingTrace
-): AnalysisResult? {
-  val solverState = get<SolverState>(SolverState.key(module))
-  return if (solverState != null && solverState.isIn(SolverState.Stage.CollectConstraints)) {
-    solverState.collectionEnds()
-    AnalysisResult.RetryWithAdditionalRoots(bindingTrace.bindingContext, module, emptyList(), emptyList())
-  } else null
-}
-
-private fun CompilerContext.ensureSolverStateInitialization(
-  module: ModuleDescriptor
-) {
-  val solverState = get<SolverState>(SolverState.key(module))
-  if (solverState == null) {
-    val state = SolverState()
-    set(SolverState.key(module), state)
-  }
-}
+import org.sosy_lab.java_smt.api.FormulaType
 
 // PHASE 2: CHECKING OF CONSTRAINTS
 // ================================
@@ -173,7 +40,7 @@ val RESULT_VAR_NAME = "${'$'}result"
 
 // 2.0: entry point
 
-private fun CompilerContext.checkDeclarationConstraints(
+internal fun CompilerContext.checkDeclarationConstraints(
   context: DeclarationCheckerContext,
   declaration: KtDeclaration,
   descriptor: DeclarationDescriptor
@@ -205,21 +72,21 @@ private fun SolverState.checkDeclarationWithBody(
   declaration: KtDeclaration,
   body: KtExpression?
 ): SimpleCont<Unit> = bracket {
-    // assert preconditions (if available)
-    checkPreconditionsInconsistencies(
-      constraints,
-      context,
-      declaration
-    ).then { inconsistentPreconditions ->
-      // if we are inconsistent, there's no point in going on, just stop early
-      guard(!inconsistentPreconditions)
-    }.then {
-      checkExpressionConstraints(resultVarName, body, context)
-    }.then {
-      // check the post-conditions
-      checkPostConditionsImplication(constraints, context, declaration)
-    }
+  // assert preconditions (if available)
+  checkPreconditionsInconsistencies(
+    constraints,
+    context,
+    declaration
+  ).then { inconsistentPreconditions ->
+    // if we are inconsistent, there's no point in going on, just stop early
+    guard(!inconsistentPreconditions)
+  }.then {
+    checkExpressionConstraints(resultVarName, body, context)
+  }.then {
+    // check the post-conditions
+    checkPostConditionsImplication(constraints, context, declaration)
   }
+}
 
 private fun SolverState.checkPreconditionsInconsistencies(
   constraints: DeclarationConstraints?,
@@ -229,7 +96,7 @@ private fun SolverState.checkPreconditionsInconsistencies(
   constraints?.pre?.let {
     addAndCheckConsistency(it) { unsatCore ->
       context.trace.report(
-        InconsistentBodyPre.on(declaration.psiOrParent, declaration, unsatCore)
+        MetaErrors.InconsistentBodyPre.on(declaration.psiOrParent, declaration, unsatCore)
       )
     }
   } ?: false // if there are no preconditions, they are consistent
@@ -243,7 +110,7 @@ private fun SolverState.checkPostConditionsImplication(
   constraints?.post?.forEach { postCondition ->
     checkImplicationOf(postCondition) {
       context.trace.report(
-        UnsatBodyPost.on(declaration.psiOrParent, declaration, listOf(postCondition))
+        MetaErrors.UnsatBodyPost.on(declaration.psiOrParent, declaration, listOf(postCondition))
       )
     }
   }
@@ -301,7 +168,7 @@ private fun SolverState.checkCallExpression(
       }
     }?.then { continueWith(it.toMap()) }
       .orElse { continueWith<Unit, Map<String, String>>(emptyMap()) }
-    // obtain and rename the pre- and post-conditions
+      // obtain and rename the pre- and post-conditions
       .then { argVars ->
         wrap {
           val callConstraints = resolvedCall?.let { constraintsFromSolverState(it) }?.let {
@@ -312,7 +179,7 @@ private fun SolverState.checkCallExpression(
           callConstraints?.pre?.forEach { callPreCondition ->
             checkImplicationOf(callPreCondition) {
               context.trace.report(
-                UnsatCallPre.on(expression.psiOrParent, resolvedCall, listOf(callPreCondition))
+                MetaErrors.UnsatCallPre.on(expression.psiOrParent, resolvedCall, listOf(callPreCondition))
               )
             }
           }
@@ -320,7 +187,7 @@ private fun SolverState.checkCallExpression(
           callConstraints?.post?.let {
             addAndCheckConsistency(it) { unsatCore ->
               context.trace.report(
-                InconsistentCallPost.on(expression.psiOrParent, resolvedCall, unsatCore)
+                MetaErrors.InconsistentCallPost.on(expression.psiOrParent, resolvedCall, unsatCore)
               )
             }
           }
@@ -371,8 +238,8 @@ private fun SolverState.checkDeclarationExpression(
   declaration: KtDeclaration,
   context: DeclarationCheckerContext
 ): SimpleCont<Unit> = declaration.stableBody()?.let {
-    checkExpressionConstraints(newVarName, it, context)
-  }.orDoNothing()
+  checkExpressionConstraints(newVarName, it, context)
+}.orDoNothing()
 
 private fun SolverState.checkNameExpression(
   associatedVarName: String,
@@ -396,34 +263,8 @@ private fun SolverState.checkNameExpression(
 
 private fun KtDeclaration.stableBody(): KtExpression?
   = when (this) {
-      is KtVariableDeclaration -> if (isVar) null else initializer
-      is KtDeclarationWithBody -> body()
-      is KtDeclarationWithInitializer -> initializer
-      else -> null
-    }
-
-// SOLVER INTERACTION
-// ==================
-// these two functions ultimately call the SMT solver,
-// and report errors as desired
-
-private fun SolverState.addAndCheckConsistency(
-  constraints: Iterable<BooleanFormula>,
-  message: (unsatCore: List<BooleanFormula>) -> Unit
-): Boolean {
-  constraints.forEach { prover.addConstraint(it) }
-  val unsat = prover.isUnsat
-  if (unsat) { message(prover.unsatCore) }
-  return unsat
+  is KtVariableDeclaration -> if (isVar) null else initializer
+  is KtDeclarationWithBody -> body()
+  is KtDeclarationWithInitializer -> initializer
+  else -> null
 }
-
-private fun SolverState.checkImplicationOf(
-  constraint: BooleanFormula,
-  message: (model: Model) -> Unit
-): Boolean =
-  bracket {
-    solver.booleans { prover.addConstraint(not(constraint)) }
-    val unsat = prover.isUnsat
-    if (!unsat) { message(prover.model) }
-    !unsat
-  }
