+++ refined-types-plugin/src/main/kotlin/arrow/meta/plugins/liquid/phases/analysis/solver/DeclarationConstraints.kt
@@ -1,4 +1,4 @@
-package arrow.meta.plugins.liquid.phases.solver.collector
+package arrow.meta.plugins.liquid.phases.analysis.solver
 
 import org.jetbrains.kotlin.descriptors.DeclarationDescriptor
 import org.jetbrains.kotlin.psi.KtDeclaration
